Move sliceviewer models and presenters to subdirectories
