+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/models/adsobsever.py
