+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/models/model.py
