+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/models/roi.py
