+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/models/transform.py
