+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/peaksviewer/representation/draw.py
@@ -11,7 +11,7 @@ from .ellipsoid import EllipsoidalIntegratedPeakRepresentation
 from .painter import MplPainter, Painted
 # 3rd party
 from mantid.geometry import PeakShape
-from mantidqt.widgets.sliceviewer.sliceinfo import SliceInfo
+from mantidqt.widgets.sliceviewer.models.sliceinfo import SliceInfo
 # standard library
 from typing import Sequence
