+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/presenters/lineplots.py
