+++ qt/python/mantidqt/mantidqt/widgets/sliceviewer/presenters/zoom.py
